Fixed Bug related to checkboxes
Bug description: When loading variant, the stored value gets discarded and the variant is overwritten with the state of the checkpoint at variant loading. This is caused because the loading algorithm loads values into the user space but checkboxes were implemented to store their values in the IN sheet.

Solution: Values are now stored in a new sheet called "Kontrollkästchen" and IN aswell as the formatting of the checkboxes is linked there.